+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add Group rename functionality (#84)
 - Add selecting fields from data frames without RefId (#85)
 - Add toggle expanded row state by clicking on the label (#86)
+- Add the deselection of All variable value which is not defined in the URL (#87)
 
 ## 2.0.0 (2023-10-16)